requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
Babel==2.9.1
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
Cython==0.29.32
Flask==2.2.2
frozenlist==1.3.1
heroku==0.1.4
idna==3.4
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==6.0.2
psycopg2==2.9.3
pyTelegramBotAPI==4.7.0
python-dateutil==1.5
python-dotenv==0.21.0
pytz==2022.2.1
requests==2.28.1
urllib3==1.26.12
vedis==0.7.1
Werkzeug==2.2.2
yarl==1.8.1
zipp==3.8.1

